# docs/swagger.yaml
definitions:
...
  models.ListeningRecordsItem:
    properties:
      answers:
        items: {}
        type: array
      id:
        type: integer
      name:
        type: string
      rest_seconds:
        type: integer
      score:
        type: integer
      status:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  models.ListeningRecordsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.ListeningRecordsItem'
        type: array
      total:
        type: integer
    type: object
...
  models.ReadingRecordsItem:
    properties:
      answers:
        items: {}
        type: array
      id:
        type: integer
      name:
        type: string
      rest_seconds:
        type: integer
      score:
        type: integer
      status:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  models.ReadingRecordsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.ReadingRecordsItem'
        type: array
      total:
        type: integer
    type: object
...
  models.ResponseData:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /record/listening/add:
    post:
      consumes:
      - application/json
      description: 新增听力做题记录
      parameters:
      - description: 听力做题记录内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.ListeningRecordsItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 新增听力做题记录
      tags:
      - Listening
  /record/listening/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除听力做题记录
      parameters:
      - description: 听力做题记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除听力做题记录
      tags:
      - Listening
  /record/listening/detail/{id}:
    get:
      consumes:
      - application/json
      description: 根据id获取听力做题记录详情
      parameters:
      - description: 听力做题记录id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ListeningRecordsItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取听力做题记录详情
      tags:
      - Listening
  /record/listening/list:
    get:
      consumes:
      - application/json
      description: 根据条件获取听力记录列表，并返回分页结果
      parameters:
      - description: 名称
        in: query
        name: name
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      - description: "3"
        in: query
        name: type
        required: true
        type: integer
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageLimit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ListeningRecordsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取听力做题记录列表
      tags:
      - Listening
  /record/listening/update:
    put:
      consumes:
      - application/json
      description: 更新听力做题记录
      parameters:
      - description: 听力做题记录内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.ListeningRecordsItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 更新听力做题记录
      tags:
      - Listening
  /record/reading/add:
    post:
      consumes:
      - application/json
      description: 新增阅读做题记录
      parameters:
      - description: 阅读做题记录内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.ReadingRecordsItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 新增阅读做题记录
      tags:
      - Reading
  /record/reading/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除阅读做题记录
      parameters:
      - description: 阅读做题记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除阅读做题记录
      tags:
      - Reading
  /record/reading/detail/{id}:
    get:
      consumes:
      - application/json
      description: 根据id获取阅读做题记录详情
      parameters:
      - description: 阅读做题记录id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ReadingRecordsItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取阅读做题记录详情
      tags:
      - Reading
  /record/reading/list:
    get:
      consumes:
      - application/json
      description: 根据条件获取阅读记录列表，并返回分页结果
      parameters:
      - description: 名称
        in: query
        name: name
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      - description: 试题类型
        in: query
        name: type
        type: integer
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageLimit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ReadingRecordsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取阅读做题记录列表
      tags:
      - Reading
  /record/reading/update:
    put:
      consumes:
      - application/json
      description: 更新阅读做题记录
      parameters:
      - description: 阅读做题记录内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.ReadingRecordsItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 更新阅读做题记录
      tags:
      - Reading
...
swagger: "2.0"